requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic-settings
passlib==1.7.4
bcrypt==4.0.1
